sctk_adwaita

Module pointer

Structs

MouseState 🔒

The state of the mouse input inside the decorations frame.

Enums

Location

Constants

DOUBLE_CLICK_DURATION 🔒

Time to register the next click as a double click.