Struct sctk_adwaita::AdwaitaFrame

pub struct AdwaitaFrame<State> {

    pub(crate) base_surface: WlTyped<WlSurface, SurfaceData>,
    pub(crate) compositor: Arc<CompositorState>,
    pub(crate) subcompositor: Arc<SubcompositorState>,
    pub(crate) queue_handle: QueueHandle<State>,
    pub(crate) decorations: Option<DecorationParts>,
    pub(crate) pool: SlotPool,
    pub(crate) dirty: bool,
    pub(crate) should_sync: bool,
    pub(crate) scale_factor: u32,
    pub(crate) resizable: bool,
    pub(crate) buttons: Buttons,
    pub(crate) state: WindowState,
    pub(crate) wm_capabilities: WindowManagerCapabilities,
    pub(crate) mouse: MouseState,
    pub(crate) theme: ColorTheme,
    pub(crate) title: Option<String>,
    pub(crate) title_text: Option<TitleText>,
    pub(crate) shadow: Shadow,

}

A simple set of decorations

Fields
base_surface: WlTyped<WlSurface, SurfaceData>

The base surface used to create the window.

compositor: Arc<CompositorState>subcompositor: Arc<SubcompositorState>

Subcompositor to create/drop subsurfaces ondemand.

queue_handle: QueueHandle<State>

Queue handle to perform object creation.

decorations: Option<DecorationParts>

The drawable decorations, None when hidden.

pool: SlotPool

Memory pool to allocate the buffers for the decorations.

dirty: bool

Whether the frame should be redrawn.

should_sync: bool

Whether the drawing should be synced with the main surface.

scale_factor: u32

Scale factor used for the surface.

resizable: bool

Wether the frame is resizable.

buttons: Buttonsstate: WindowStatewm_capabilities: WindowManagerCapabilitiesmouse: MouseStatetheme: ColorThemetitle: Option<String>title_text: Option<TitleText>shadow: Shadow
Implementations

impl<State> AdwaitaFrame<State>
where State: Dispatch<WlSurface, SurfaceData> + Dispatch<WlSubsurface, SubsurfaceData> + 'static,

pub fn new( base_surface: &impl WaylandSurface, shm: &Shm, compositor: Arc<CompositorState>, subcompositor: Arc<SubcompositorState>, queue_handle: QueueHandle<State>, frame_config: FrameConfig, ) -> Result<Self, Box<dyn Error>>

pub fn set_config(&mut self, config: FrameConfig)

Update the current frame config.

pub(crate) fn precise_location( &self, location: Location, decoration: &DecorationParts, x: f64, y: f64, ) -> Location

pub(crate) fn redraw_inner(&mut self) -> Option<bool>