sctk_adwaita::theme

Struct Color

pub struct Color { /* private fields */ }

An RGBA color value, holding four floating point components.

Guarantees

All values are in 0..=1 range.

Implementations

impl Color

pub const TRANSPARENT: Color

A transparent color.

pub const BLACK: Color

A black color.

pub const WHITE: Color

A white color.

pub fn from_rgba(r: f32, g: f32, b: f32, a: f32) -> Option<Color>

Creates a new color from 4 components.

All values must be in 0..=1 range.

pub fn from_rgba8(r: u8, g: u8, b: u8, a: u8) -> Color

Creates a new color from 4 components.

u8 will be divided by 255 to get the float component.

pub fn red(&self) -> f32

Returns color’s red component.

The value is guarantee to be in a 0..=1 range.

pub fn green(&self) -> f32

Returns color’s green component.

The value is guarantee to be in a 0..=1 range.

pub fn blue(&self) -> f32

Returns color’s blue component.

The value is guarantee to be in a 0..=1 range.

pub fn alpha(&self) -> f32

Returns color’s alpha component.

The value is guarantee to be in a 0..=1 range.

pub fn set_red(&mut self, c: f32)

Sets the red component value.

The new value will be clipped to the 0..=1 range.

pub fn set_green(&mut self, c: f32)

Sets the green component value.

The new value will be clipped to the 0..=1 range.

pub fn set_blue(&mut self, c: f32)

Sets the blue component value.

The new value will be clipped to the 0..=1 range.

pub fn set_alpha(&mut self, c: f32)

Sets the alpha component value.

The new value will be clipped to the 0..=1 range.

pub fn apply_opacity(&mut self, opacity: f32)

Shifts color’s opacity.

Essentially, multiplies color’s alpha by opacity.

opacity will be clamped to the 0..=1 range first. The final alpha will also be clamped.

pub fn is_opaque(&self) -> bool

Check that color is opaque.

Alpha == 1.0

pub fn premultiply(&self) -> PremultipliedColor

Converts into a premultiplied color.

pub fn to_color_u8(&self) -> ColorU8

Converts into ColorU8.